+++ db/revert/tables/application_form_template_9_data.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:tables/application_form_template_9_data from pg
+
+BEGIN;
+
+drop table if exists ccbc_public.application_form_template_9_data;
+
+COMMIT;